refactor(printer): split printer.rs into module directory (closes #247)
Mechanical split of the 1,662-line printer.rs into six topical files
under printer/ for easier navigation. No logic changes; lib.rs still
re-exports Indent/PrinterConfig so the public API is unchanged.

Layout:
- mod.rs    — PrinterConfig, Indent, Printer struct, write helpers,
              print_program, print_stmts
- stmts.rs  — print_stmt dispatch, print_if, print_try_catch,
              print_block_or_stmt, print_namespace, print_use
- decls.rs  — print_function, print_class, print_method, print_property,
              property hooks, traits, enums, params, attributes,
              doc comments
- exprs.rs  — print_expr dispatch, closures, arrow functions, match,
              args, array elements, string parts/literals
- types.rs  — print_name, print_type_hint
- helpers.rs — free helpers (op strings, escapes, is_declaration)

Also adds 17 new printer fixtures to cover previously-untested edge
cases: heredoc/nowdoc, shell_exec, declare (with and without body),
trait-use adaptations, match default arm, property hooks (PHP 8.4),
throw expression, dynamic class const access, magic constants
NAMESPACE/TRAIT/PROPERTY, halt_compiler, inline HTML, empty class
bodies, anonymous class with constructor args, dynamic static property
access, and braced variable variable.
