+++ crates/php-printer/tests/fixtures/anonymous_class_with_args.phpt
@@ -1,0 +1,12 @@
+===source===
+<?php $obj = new class("a", 1) extends Base implements I1, I2 {
+    public function f() { return 1; }
+};
+===print===
+$obj = new class extends Base implements I1, I2
+{
+    public function f()
+    {
+        return 1;
+    }
+}('a', 1);
