+++ crates/php-printer/tests/fixtures/declare_simple.phpt
@@ -1,0 +1,4 @@
+===source===
+<?php declare(strict_types=1);
+===print===
+declare(strict_types=1);
