+++ crates/php-printer/tests/fixtures/declare_with_body.phpt
@@ -1,0 +1,8 @@
+===source===
+<?php declare(ticks=1) {
+    echo "in block";
+}
+===print===
+declare(ticks=1) {
+    echo 'in block';
+}
