+++ crates/php-printer/tests/fixtures/dynamic_class_const_access.phpt
@@ -1,0 +1,4 @@
+===source===
+<?php $v = Foo::{$name};
+===print===
+$v = Foo::{$name};
