+++ crates/php-printer/tests/fixtures/empty_class_body.phpt
@@ -1,0 +1,17 @@
+===source===
+<?php class Empty1 {}
+interface Empty2 {}
+trait Empty3 {}
+enum Empty4 {}
+===print===
+class Empty1
+{}
+
+interface Empty2
+{}
+
+trait Empty3
+{}
+
+enum Empty4
+{}
