+++ crates/php-printer/tests/fixtures/halt_compiler.phpt
@@ -1,0 +1,7 @@
+===source===
+<?php echo "before";
+__halt_compiler();
+trailing data here
+===print===
+echo 'before';
+__halt_compiler();trailing data here
