+++ crates/php-printer/tests/fixtures/heredoc.phpt
@@ -1,0 +1,9 @@
+===source===
+<?php $x = <<<EOT
+hello $name
+world
+EOT;
+===print===
+$x = <<<EOT
+hello $name\nworld
+EOT;
