+++ crates/php-printer/tests/fixtures/inline_html.phpt
@@ -1,0 +1,10 @@
+===source===
+<?php if ($cond) { ?>
+<div>hello</div>
+<?php } ?>
+===print===
+if ($cond) {
+    ?>
+<div>hello</div>
+<?php
+}
