+++ crates/php-printer/tests/fixtures/magic_constants_extra.phpt
@@ -1,0 +1,8 @@
+===source===
+<?php $a = __NAMESPACE__;
+$b = __TRAIT__;
+$c = __PROPERTY__;
+===print===
+$a = __NAMESPACE__;
+$b = __TRAIT__;
+$c = __PROPERTY__;
