+++ crates/php-printer/tests/fixtures/match_with_default.phpt
@@ -1,0 +1,12 @@
+===source===
+<?php $r = match ($x) {
+    1, 2 => "low",
+    3 => "mid",
+    default => "other",
+};
+===print===
+$r = match ($x) {
+    1, 2 => 'low',
+    3 => 'mid',
+    default => 'other',
+};
