+++ crates/php-printer/tests/fixtures/nowdoc.phpt
@@ -1,0 +1,10 @@
+===source===
+<?php $x = <<<'EOT'
+literal $not_interpolated
+text
+EOT;
+===print===
+$x = <<<'EOT'
+literal $not_interpolated
+text
+EOT;
