+++ crates/php-printer/tests/fixtures/property_hooks.phpt
@@ -1,0 +1,26 @@
+===source===
+<?php class User {
+    public string $name {
+        get => strtoupper($this->name);
+        set(string $value) {
+            $this->name = trim($value);
+        }
+    }
+
+    public string $email { get; set; }
+}
+===print===
+class User
+{
+    public string $name {
+        get => strtoupper($this->name);
+        set(string $value) {
+            $this->name = trim($value);
+        }
+    }
+
+    public string $email {
+        get;
+        set;
+    }
+}
