+++ crates/php-printer/tests/fixtures/shell_exec.phpt
@@ -1,0 +1,4 @@
+===source===
+<?php $out = `ls -la $dir`;
+===print===
+$out = `ls -la $dir`;
