+++ crates/php-printer/tests/fixtures/static_property_access_dynamic.phpt
@@ -1,0 +1,4 @@
+===source===
+<?php $v = Foo::$$name;
+===print===
+$v = Foo::$$name;
