+++ crates/php-printer/tests/fixtures/throw_expression.phpt
@@ -1,0 +1,4 @@
+===source===
+<?php $x = $value ?? throw new Exception("missing");
+===print===
+$x = $value ?? (throw new Exception('missing'));
