+++ crates/php-printer/tests/fixtures/trait_use_adaptations.phpt
@@ -1,0 +1,19 @@
+===source===
+<?php class C {
+    use A, B {
+        A::foo insteadof B;
+        B::foo as barFromB;
+        B::bar as protected;
+        B::baz as protected renamedBaz;
+    }
+}
+===print===
+class C
+{
+    use A, B {
+        A::foo insteadof B;
+        B::foo as barFromB;
+        B::bar as protected;
+        B::baz as protected renamedBaz;
+    }
+}
