+++ crates/php-printer/tests/fixtures/variable_variable_braced.phpt
@@ -1,0 +1,4 @@
+===source===
+<?php $a = ${'foo' . $i};
+===print===
+$a = ${'foo' . $i};
